finished serialization for GDLevel
tests are still not passing though
